Making sure the usage is clear for the Thor tasks.
tasks.rb represents the actual Thor tasks. They can be included in other gems.
Change cli.rb to simply include the tasks
Add the repository configuration option to the readme,
