+++ lib/nexus_cli.rb
@@ -1,7 +1,8 @@
-require "nexus_cli/remote"
-require "nexus_cli/cli"
-require "nexus_cli/errors"
-require "nexus_cli/kernel"
+require 'nexus_cli/tasks'
+require 'nexus_cli/cli'
+require 'nexus_cli/errors'
+require 'nexus_cli/kernel'
+require 'nexus_cli/remote'
 
 module NexusCli
   class << self
